Refactor out some program-global variables
Removed most of the program-global in_addr variables; most have moved to
become module-global in static.c, but ipaddress is fully local now, and is
passed around where it needs to be (as a proof-of-concept for the other
variables at a later time).

The calculation of network and broadcast addresses was rearranged somewhat
as a consequence; it has been placed closer to where it's actually used, and
abstracted into a couple of helper functions.

Tested with d-i runs through to mirror selection with DHCP on, DHCP off
(static address config), and a PLIP (pointopoint) install (add plip-modules
to the initrd and modprobe plip before doing anything -- works great).