Remove gateway from being a global var, and turn it into a string
The 'gateway' global variable has been completely removed, and turned into a
local variable to be passed around as needed.

I also, as the major part of this patch, changed the type of the variable
from a struct in_addr into a char *.  This has two advantages:

* Less inet_{pton,ntop} calls: At the moment, there are still a lot of them
  because other network data variables are still in_addrs, but most of the
  time we need all of those variables as strings.
* IPv6 compatibility: Storing IP addresses as strings makes IPv6
  compatibility a lot easier; rather than needing unions and flag variables,
  we can just store everything in strings.  Yippee!

All other relevant variables will be converted into strings down the line,
too.

Rebased-and-modified-by: Philipp Kern <pkern@debian.org>