Create and use a struct to store all network interface settings
Please welcome struct netcfg_interface.  It's a place to keep all of the
fundamental configuration about an interface -- device name, IP address,
gateway, netmask, etc.  This results in the removal of the 'interface'
global, which just kept the interface name, as well as the renaming of a lot
of places that used 'interface' to if_name, to make it obvious that we're
talking about the name of an interface, not all of the details about it.

Rebased-and-modified-by: Philipp Kern <pkern@debian.org>