Refactor /etc/network/interfaces writing code
Remove duplication, centralise all this common functionality rather than
leave it scattered around everywhere, and robustify the code in the face of
potential (though, let's face it, very unlikely) errors.

The result of this is much improved clarity and extensibility (important, in
light of planned changes), and -- for now -- the addition of SLAAC support
code.

Possibly-broken-by: Philipp Kern <pkern@debian.org>

Conflicts:
	dhcp.c
	netcfg-common.c