Use dhclient -6 to support stateless DHCP on FreeBSD
wide-dhcpv6 (dhcp6c) is nice and tiny, but is Linux only.  dhclient is
cross-platform, but is 3x the size of udhcpc+dhcp6c combined, so we'd prefer
not to use it everywhere.

Thankfully, with a bit of hacking we can have dhclient act like dhcp6c, and
produce a useful stateless config on kFreeBSD.

Rebased-and-modified-by: Philipp Kern <pkern@debian.org>