Remove global variable netcfg_progress_displayed
It was only used to guard a call to debconf_progress_stop() in netcfg_die();
I've checked a few debconf frontend implementations, and they all seem to
handle a PROGRESS STOP message when they're not in a progress bar in a sane
and safe manner.  I've also tested that there's no ill effects in practice
using the default (newt) frontend.