ipv6.c: force stateless_config to 0 if stateful_config is set also
If stateful_config is set, we need to keep the DHCPv6 client running and
not only retrieve configuration but request addresses too. Hence set
stateless_config to 0 if both managed and other are set in Router
Advertisements. This case is currently underdefined but we want stateful
configuration to happen in any case.

Signed-off-by: Philipp Kern <pkern@debian.org>