Remove DHCPv6 DUID-LL generation.
DUID-LL is specified in RFC 3315 with the following note:

   DUID-LL is recommended for devices that have a permanently-connected
   network interface with a link-layer address, and do not have
   nonvolatile, writable stable storage.  DUID-LL MUST NOT be used by
   DHCP clients or servers that cannot tell whether or not a network
   interface is permanently attached to the device on which the DHCP
   client is running.

As much as I have empathy with the author of that stanza to use DUID-LL
for predictable IDs, it would be possible if the card is taken out of
the box and used in another machine for installation to get the same
DUID. As far as I can see it would only be used for installation and not
be copied into the installed system, too. (Same for DUID-LLT.) DHCPv6
does use DUID-LLT everywhere and diverging in netcfg does not make much
sense, especially if we violate the RFC doing that.

If you want to use a predictable MAC address-based scheme, please use
IPv6 stateless autoconfiguration which does exactly this.

Signed-off-by: Philipp Kern <pkern@debian.org>