Reword target_network_config template; do not ask the question.
While netcfg/target_network_config might be useful for preseeding it
generally is not for regular use of the installer. Especially not
with medium priority. Hence use the heuristics that n-m configuration
is used when n-m got installed and ifupdown otherwise, with wireless
not being configured in /e/n/i if present.

Signed-off-by: Philipp Kern <pkern@debian.org>