nm-conf: Allow both IPv4 and IPv6 to be activated post-installation.
Align the write out of IPv4 and IPv6 to the defaults in network-manager.
A connection is considered up if one of IPv4 and IPv6 is up. Allow
each of IPv4 and IPv6 to be activated post-installation by specifying
"auto" as their configuration method, even if they were not present.
Always write out IPv4 and IPv6 configuration blocks like nm-applet does.

Signed-off-by: Philipp Kern <pkern@debian.org>