nm-conf: Only write out MAC addresses on static configurations.
Specifying the MAC address in network-manager's configuration files
means binding that particular configuration to a specific device. This
might be unhelpful, especially if the configuration is nothing more than
network-manager's default anyway. However in case of a static
configuration it does make sense to bind it to the device which was used
for the installation.  Hence retain the MAC write out for this.

Signed-off-by: Philipp Kern <pkern@debian.org>